backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
httpx-sse==0.4.0
aiomysql==0.2.0
sqlalchemy[asyncio]==2.0.30
redis[hiredis]==5.0.0
pyjwt==2.9.0
pydantic==2.7.0
pydantic-settings==2.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
